requirements.txt
python-telegram-bot==20.8
sqlite3-binary

